+++ src/utils/render.js
@@ -1,0 +1,35 @@
+import AbstractView from '../view/abstract-view.js';
+
+export const RenderPosition = {
+  BEFORE_BEGIN: 'beforebegin',
+  AFTER_BEGIN: 'afterbegin',
+  BEFORE_END: 'beforeend',
+  AFTER_END: 'afterend',
+};
+
+export const renderElement = (container, element) => {
+  const parent = container instanceof AbstractView ? container.element : container;
+  const child = element instanceof AbstractView ? element.element : element;
+
+  parent.append(child);
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement('div');
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
+
+
+export const removeComponent = (component) => {
+  if (component === null) {
+    return;
+  }
+
+  if(!(component instanceof AbstractView)) {
+    throw new Error('Can remove only components');
+  }
+
+  component.element.remove();
+  component.removeElement();
+};
