+++ src/view/abstract-view.js
@@ -1,0 +1,28 @@
+import { createElement } from '../utils/render.js';
+
+export default class AbstractView {
+  #element = null;
+  _callback = {};
+
+  constructor () {
+    if (new.target === AbstractView) {
+      throw new Error('Can\'t instantiate AbstractView, only concrete one.');
+    }
+  }
+
+  get element () {
+    if(!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  get template () {
+    throw new Error('Abstract method not implemented: get template');
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
