+++ src/view/button-show-more-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createButtonShowMoreTemplate = () => (
   `<button class="films-list__show-more">
@@ -6,22 +6,18 @@ const createButtonShowMoreTemplate = () => (
   </button>`
 );
 
-export default class ButtonShowMoreView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ButtonShowMoreView extends AbstractView {
   get template() {
     return createButtonShowMoreTemplate();
   }
 
-  removeElement() {
-    this.#element = null;
+  setClickHandler = (callback) => {
+    this._callback.click = callback;
+    this.element.addEventListener('click', this.#clickHandler);
+  }
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
   }
 }
