+++ src/view/film-section-view.js
@@ -1,25 +1,11 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmSectionTemplate = () => (
   '<section class="films"></section>'
 );
 
-export default class FilmSectionView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmSectionView extends AbstractView {
   get template() {
     return createFilmSectionTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
