+++ src/view/films-list-container-view.js
@@ -1,25 +1,11 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmsListContainerTemplate = () => (
   '<div class="films-list__container"></div>'
 );
 
-export default class FilmsListContainerView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilmsListContainerView extends AbstractView {
   get template() {
     return createFilmsListContainerTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
