+++ src/view/menu-container-view.js
@@ -1,26 +1,12 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createMenuContainerTemplate = () => (
   `<nav class="main-navigation">
   </nav>`
 );
 
-export default class MenuContainerView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class MenuContainerView extends AbstractView {
   get template() {
     return createMenuContainerTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
