+++ src/view/menu-stats-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createMenuStatsTemplate = () => (
   `<a href="#stats" class="main-navigation__additional">
@@ -6,22 +6,8 @@ const createMenuStatsTemplate = () => (
   </a>`
 );
 
-export default class MenuStatsView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class MenuStatsView extends AbstractView {
   get template() {
     return createMenuStatsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
