+++ src/view/popup-close-button-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createPopupCloseButtonTemplate = () => (
   `<button class="film-details__close-btn" type="button">
@@ -6,22 +6,18 @@ const createPopupCloseButtonTemplate = () => (
   </button>`
 );
 
-export default class PopupCloseButtonView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class PopupCloseButtonView extends AbstractView {
   get template() {
     return createPopupCloseButtonTemplate();
   }
 
-  removeElement() {
-    this.#element = null;
+  setClickHandler = (callback) => {
+    this._callback.click = callback;
+    this.element.addEventListener('click', this.#clickHandler);
+  }
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
   }
 }
