+++ src/view/popup-comment-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createPopupCommentTemplate = (comment) => (
   `<li class="film-details__comment">
@@ -16,27 +16,15 @@ const createPopupCommentTemplate = (comment) => (
   </li>`
 );
 
-export default class PopupCommentView {
-  #element = null;
+export default class PopupCommentView extends AbstractView {
   #comment = null;
 
   constructor(comment) {
+    super();
     this.#comment = comment;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createPopupCommentTemplate(this.#comment);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
