+++ src/view/popup-emoji-list-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createPopupEmojiListTemplate = (emojiList, isChecked) => (
   `<div class="film-details__emoji-list">
@@ -10,29 +10,17 @@ const createPopupEmojiListTemplate = (emojiList, isChecked) => (
   </div>`
 );
 
-export default class PopupEmojiListView {
-  #element = null;
+export default class PopupEmojiListView extends AbstractView {
   #emojiList = null;
   #isChecked = null;
 
   constructor(emojiList, isChecked) {
+    super();
     this.#emojiList = emojiList;
     this.#isChecked = isChecked;
   }
 
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
   get template() {
     return createPopupEmojiListTemplate(this.#emojiList, this.#isChecked);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
